+++ src/main/kotlin/com/dicoding/exam/latihanopsional1/App.kt
@@ -10,27 +10,26 @@
  *                                                                                                  *
  ***************************************************************************************************/
 
-package com.dicoding.exam.latihan_opsional4
+package com.dicoding.exam.latihanopsional1
 
 /**
  *  TODO
  *  Lengkapi fungsi di bawah dengan ketentuan sebagai berikut:
- *  - Fungsi menerima input string dengan panjang karakter 1-100.
- *  - Fungsi harus mengembalikan karakter yang ditengah dari string yang diinput.
- *      - Jika panjang string bernilai genap, maka yang dikembalikan adalah 2 karakter.
- *      - Jika panjang string bernilai ganjil, maka yang dikembalikan adalah 1 karakter.
+ *  - Fungsi menerima input string berupa kode warna dalam format 6 digit hexadesimal. Contoh: #FFFFFF atau #OOOOOO.
+ *  - Fungsi mengembalikan class RBG dengan parameter r untuk merah, g untuk hijau dan b untuk biru.
+ *    Dimana masing-masing paramter tersebut bertipe Interger dengan rentang nilai 0-255.
  *
  *  Contoh:
- *  Input = string=dicoding -> Output = od
- *  Input = string=dicodingindonesia  -> Output = i
- *  Input = string=A  -> Output = A
+ *  Input = #FF2231 -> Output = RGB(r=255, g=34, b=49)
+ *  Input = #123456 -> Output = RGB(r=18, g=52, b=86)
  */
 
-fun getMiddleCharacters(string: String): String {
-    return ""
+data class RGB(val r: Int, val g: Int, val b: Int)
+
+fun hexColorToRGB(color: String): RGB {
+    return RGB(0, 0, 0)
 }
 
 fun main() {
-    println(getMiddleCharacters("dicodingindonesia") == "i")
-    println(getMiddleCharacters("dicoding") == "od")
+    println("${hexColorToRGB("#123456") == RGB(18, 52, 86)}")
 }
